Restructure folders and files to follow best practices. Create (auth) and (main) folders since they have different layouts and avoid conflicts. Change styles on product cards and colors on main page. Add background image to main page.
