+++ app/(main)/dashboard/page.tsx
@@ -1,0 +1,8 @@
+import ProductCard from "@/components/ProductCard/ProductCard";
+
+export default function Home() {
+    return (
+        <main className="">
+            <ProductCard />
+        </main>)
+}
