+++ app/(main)/purchases/page.tsx
