+++ app/styles/global.css
@@ -10,7 +10,7 @@
 .page {
   display: flex;
   align-items: center;
-  background-color: #f6e9d5;
+  background-image: url('../../public/bckgrnd-img.png');
   min-height: 100vh;
 }
 
@@ -18,6 +18,7 @@
   width: 100%;
   min-height: 95vh;
   display: grid;
+  background-color: rgba(255, 255, 255, 0.612);
   grid-template-rows: minmax(72px, 80px) auto minmax(20px, 40px);
   max-width: 1360px;
   margin: 0 auto;
