+++ auth.ts
@@ -1,39 +1,41 @@
-import NextAuth from 'next-auth';
-import Credentials from 'next-auth/providers/credentials';
-import { LoginFormSchema } from '@/app/lib/schemas';
-import { authConfig } from './auth.config';
-import type { User } from '@/app/lib/definitions';
-import bcrypt from 'bcrypt';
-import postgres from 'postgres';
- 
-const sql = postgres(process.env.POSTGRES_URL!, { ssl: 'require' });
- 
+import NextAuth from "next-auth";
+import Credentials from "next-auth/providers/credentials";
+import { LoginFormSchema } from "@/app/(main)/lib/schemas";
+import { authConfig } from "./auth.config";
+import type { User } from "@/app/(main)/lib/definitions";
+import bcrypt from "bcrypt";
+import postgres from "postgres";
+
+const sql = postgres(process.env.POSTGRES_URL!, { ssl: "require" });
+
 async function getUser(email: string): Promise<User | undefined> {
   try {
     const user = await sql<User[]>`SELECT * FROM users WHERE email=${email}`;
     return user[0];
   } catch (error) {
-    console.error('Failed to fetch user:', error);
-    throw new Error('Failed to fetch user.');
+    console.error("Failed to fetch user:", error);
+    throw new Error("Failed to fetch user.");
   }
 }
- 
+
 export const { auth, signIn, signOut } = NextAuth({
   ...authConfig,
   providers: [
     Credentials({
       async authorize(credentials) {
-        const parsedCredentials = LoginFormSchema
-          .safeParse(credentials);
- 
+        const parsedCredentials = LoginFormSchema.safeParse(credentials);
+
         if (parsedCredentials.success) {
           const { email, password } = parsedCredentials.data;
           const user = await getUser(email);
           if (!user) return null;
-          const passwordsMatch = await bcrypt.compare(password, user.password_hash);
+          const passwordsMatch = await bcrypt.compare(
+            password,
+            user.password_hash,
+          );
           if (passwordsMatch) return user;
         }
-        console.log('Invalid credentials provided.');
+        console.log("Invalid credentials provided.");
         return null;
       },
     }),
