+++ components/header/Header.css
@@ -1,23 +1,22 @@
 .header {
-    background-color: #a7c957;
+    background-color: rgb(41, 41, 41);
     border-radius: 10px;
     width: 100%;
     display: flex;
     align-items: center;
+    color: #fff;
 }
 
 .fa-house {
-    color: #2B2E37;
     font-size: 2rem;
 }
 
 .fa-leaf {
-    color: #fff;
+    color: #ffd92f;
     font-size: 1.5rem;
 }
 
 .header__title {
     font-size: 1.5rem;
     font-weight: bold;
-    color: #2B2E37;
 }
